refactor: typing cleanup
