+++ src/app/about/status/page.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import ActiveIssues from "@/components/status/ActiveIssues"
 import IssueGrid from "@/components/status/IssueGrid"
-import { getOpenIssues } from "@/lib/get-open-issues"
+import { getOpenIssues } from "@/lib/fetch-issues"
 
 export default async function ActiveStatus() {
   const issues = await getOpenIssues()
